requirements.txt
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
awkward==2.2.4
awkward-cpp==17
awkward0==0.15.5
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
cachetools==5.5.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.2
fastjsonschema==2.20.0
fonttools==4.38.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.6
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.2.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.6
traitlets==5.9.0
typing_extensions==4.7.1
uproot==5.0.13
uproot3==3.14.4
uproot3-methods==0.10.1
vector==1.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
zipp==3.15.0
zstandard==0.21.0
